requirements.txt
argparse
requests
pillow
beautifulsoup4
